Breakout EllipseDrawable
